+++ cartridges/int_marketpay_headless/cartridge/rest-apis/marketpay/script.js
@@ -1,15 +1,11 @@
 const RESTResponseMgr = require('dw/system/RESTResponseMgr');
 const marketPay = require('*/cartridge/scripts/services/marketPay')
 const Logger = require('dw/system/Logger').getLogger('MarketPay','MarketPay');
-const SCAPIService = require('*/cartridge/scripts/services/scapiService');
 
 exports.createCheckoutSession = function () {
-    var customerId = request.httpParameterMap.c_customerId;
-    var customerIdValue = customerId.stringValue;
-
     try {
-        var requestBody = request.httpParameterMap.requestBodyAsString;
-        var requestData = JSON.parse(requestBody);
+        var customerId = request.httpParameterMap.c_customerId.stringValue;
+        var requestData = JSON.parse(request.httpParameterMap.requestBodyAsString);
         var result = marketPay.getTokenAndSessionId(requestData);
         var paymentMethods = marketPay.getPaymentMethods(result.token, result.sessionId);
 
@@ -21,12 +17,12 @@ exports.createCheckoutSession = function () {
         var CustomObjectMgr = require('dw/object/CustomObjectMgr');
 
         Transaction.wrap(function () {
-            var marketPayDataObj = CustomObjectMgr.getCustomObject('MarketPayData', customerIdValue);
+            var marketPayDataObj = CustomObjectMgr.getCustomObject('MarketPayData', customerId);
             if (!marketPayDataObj) {
-                marketPayDataObj = CustomObjectMgr.createCustomObject('MarketPayData', customerIdValue);
+                marketPayDataObj = CustomObjectMgr.createCustomObject('MarketPayData', customerId);
             }
             marketPayDataObj.custom.sessionID = result.sessionId;
-            marketPayDataObj.custom.customerID = customerIdValue;
+            marketPayDataObj.custom.customerID = customerId;
             marketPayDataObj.custom.token = result.token;
             marketPayDataObj.custom.paymentMethods = JSON.stringify(paymentMethods);
         });        
@@ -35,14 +31,10 @@ exports.createCheckoutSession = function () {
             .createEmptySuccess(200)
             .render();
     } catch (error) {
-
         Logger.error('Error creating session ' + error.message);
-
         RESTResponseMgr
             .createError(404, 'Session-error', 'Not created', 'please reach out the SFCC developers.')
             .render();
-
-        
     }
 };